fix(096-W4-4-3): 修评审——空 catch 补 warn + 主组件再拆两表（469→43 行）+ Playwright 补真实交互
回应双轴评审：

【Standards】
- 禁止清单 #6 空 catch：RunConfigCard.loadConfig / UsageStatsCard.loadUsageStatsSettings /
  useExecutorAdmin.fetchExecutorModels 三处 catch 补 console.warn（带原因）。
- 禁止清单 #14 单组件 >300 行 + 函数 >50 行：主组件再下沉两张表——
  ExecutorsTable（执行器列表+列交互+测试 Modal+两配置卡片）、RunningRecordsTable（运行中记录表），
  各列 render 拆为 ≤50 行命名 helper（renderModelSelect/renderActions/renderInstallButton...）。
  主组件 ExecutorsPanel 469→43 行（纯 Tabs 编排），ExecutorsTable 288、RunConfigCard 161、UsageStatsCard 112 均 <300。

【Spec】
- 101 验证门禁「行内编辑保存/检测按钮/默认切换」此前仅渲染检查：补 3 个真实交互用例——
  检测点击→状态落结果（detectExecutorByName 端到端）、二进制路径编辑→跨重载持久（saveExecutorField）、
  设为默认翻转+还原（setAsDefault 全表重算 is_default）。均带 dev 配置还原。

验证：tsc 零错误 / vitest 402 全过 / build 通过 / Playwright 7/7（4 渲染 + 3 交互）。
